+++ src/gql/base.resolver.ts
@@ -11,8 +11,9 @@ import {
 } from 'type-graphql';
 import { Oid } from '@rumbleship/oid';
 import { Scopes } from '@rumbleship/acl';
+// tslint:disable-next-line: no-circular-imports
 import { RelayService, Node, Connection } from './index';
-import { Context } from './../server/context.interface';
+import { RumbleshipContext } from './../server/context.interface';
 import { ClassType } from './../helpers/classtype';
 import { NodeNotification, NODE_CHANGE_NOTIFICATION } from './node-notification';
 import { createPayloadUsingStr, RawPayload } from '../pubsub/helper';
@@ -24,7 +25,7 @@ export class GQLBaseResolver<
   TInput,
   TUpdate
 > {
-  public ctx: Context;
+  public ctx: RumbleshipContext;
   constructor(protected service: RelayService<TApi, TConnection, TFilter, TInput, TUpdate>) {
     this.ctx = service.getContext();
     service.nodeType();